oficios-automation/funcoes/W0_gmail_ingest/requirements.txt
functions-framework==3.*
google-api-python-client==2.*
google-auth==2.*
google-auth-httplib2==0.*
google-auth-oauthlib==1.*
google-cloud-secret-manager==2.*
google-cloud-storage==2.*


